+++ app/thankyou/page.tsx
@@ -2,6 +2,6 @@ import ThankyouPage from ".";
 
 export const dynamicParams = false;
 
-export default async function NetUG() {
+export default async function Thankyou() {
   return <ThankyouPage />;
 }
